requirements.txt

fastapi
uvicorn
python-dotenv
pydantic
requests
streamlit

langchain
langchain-openai
langchain-community
langchain-core

pymongo[srv]
>>>>>>> dd77fef (Your update message)
